+++ levelsetfield.h
@@ -13,7 +13,7 @@ public:
     virtual void clear();
     virtual double getFieldValue(glm::vec3 p);
 
-    void setSignedDistanceField(Array3d<double> distField);
+    void setSignedDistanceField(Array3d<float> distField);
     bool isCellInside(GridIndex g) { return _distanceField(g) > 0.0; };
     double getFieldValueAtCellCenter(GridIndex g) { return _distanceField(g); };
 
@@ -22,6 +22,6 @@ private:
     double _fasttricubicInterpolate(double p[4][4][4], double x, double y, double z);
 
     double _surfaceThreshold = 0.0;
-    Array3d<double> _distanceField;
+    Array3d<float> _distanceField;
 };
